use cramer's rule

+++ 2024/Day13/Day13.UnitTests/ClawContraptionsMust.cs
@@ -20,23 +20,23 @@ public class ClawContraptionsMust
         var sut = new ClawContraptions(SAMPLE_INPUT);
         Assert.Collection(sut.Contraptions,
             p1 => {
-                Assert.Equal((94, 34), p1.ButtonA);
-                Assert.Equal((22, 67), p1.ButtonB);
+                Assert.Equal((94UL, 34UL), p1.ButtonA);
+                Assert.Equal((22UL, 67UL), p1.ButtonB);
                 Assert.Equal((8400UL, 5400UL), p1.Prize);
             },
             p2 => {
-                Assert.Equal((26, 66), p2.ButtonA);
-                Assert.Equal((67, 21), p2.ButtonB);
+                Assert.Equal((26UL, 66UL), p2.ButtonA);
+                Assert.Equal((67UL, 21UL), p2.ButtonB);
                 Assert.Equal((12748UL, 12176UL), p2.Prize);
             },
             p3 => {
-                Assert.Equal((17, 86), p3.ButtonA);
-                Assert.Equal((84, 37), p3.ButtonB);
+                Assert.Equal((17UL, 86UL), p3.ButtonA);
+                Assert.Equal((84UL, 37UL), p3.ButtonB);
                 Assert.Equal((7870UL, 6450UL), p3.Prize);
             },
             p1 => {
-                Assert.Equal((69, 23), p1.ButtonA);
-                Assert.Equal((27, 71), p1.ButtonB);
+                Assert.Equal((69UL, 23UL), p1.ButtonA);
+                Assert.Equal((27UL, 71UL), p1.ButtonB);
                 Assert.Equal((18641UL, 10279UL), p1.Prize);
             });
     }
@@ -44,7 +44,7 @@ public class ClawContraptionsMust
     [Theory]
     [InlineData("Button A: X+94, Y+34\nButton B: X+22, Y+67\nPrize: X=8400, Y=5400", 280)]
     [InlineData("Button A: X+17, Y+86\nButton B: X+84, Y+37\nPrize: X=7870, Y=6450", 200)]
-    public void FindBestCombination(string input, int expectedWin)
+    public void FindBestCombination(string input, ulong expectedWin)
     {
         var sut = new ClawContraptions(input);
         Assert.Equal(expectedWin, sut.CheapestWin);
@@ -56,21 +56,21 @@ public class ClawContraptionsMust
     public void ReturnZero_WhenNoCombinationHasBeenFound(string input)
     {
         var sut = new ClawContraptions(input);
-        Assert.Equal(0, sut.CheapestWin);
+        Assert.Equal(0UL, sut.CheapestWin);
     }
 
     [Fact]
     public void SolveFirstSampleCorrectly()
     {
         var sut = new ClawContraptions(SAMPLE_INPUT);
-        Assert.Equal(480, sut.CheapestWin);
+        Assert.Equal(480UL, sut.CheapestWin);
     }
 
     [Fact]
     public void SolveFirstPuzzleCorrectly()
     {
         var sut = new ClawContraptions(PUZZLE_INPUT);
-        Assert.Equal(28059, sut.CheapestWin);
+        Assert.Equal(28059UL, sut.CheapestWin);
     }
 
     [Fact]
@@ -79,23 +79,23 @@ public class ClawContraptionsMust
         var sut = new ClawContraptions(SAMPLE_INPUT, 10000000000000);
         Assert.Collection(sut.Contraptions,
             p1 => {
-                Assert.Equal((94, 34), p1.ButtonA);
-                Assert.Equal((22, 67), p1.ButtonB);
+                Assert.Equal((94UL, 34UL), p1.ButtonA);
+                Assert.Equal((22UL, 67UL), p1.ButtonB);
                 Assert.Equal((10000000008400UL, 10000000005400UL), p1.Prize);
             },
             p2 => {
-                Assert.Equal((26, 66), p2.ButtonA);
-                Assert.Equal((67, 21), p2.ButtonB);
+                Assert.Equal((26UL, 66UL), p2.ButtonA);
+                Assert.Equal((67UL, 21UL), p2.ButtonB);
                 Assert.Equal((10000000012748UL, 10000000012176UL), p2.Prize);
             },
             p3 => {
-                Assert.Equal((17, 86), p3.ButtonA);
-                Assert.Equal((84, 37), p3.ButtonB);
+                Assert.Equal((17UL, 86UL), p3.ButtonA);
+                Assert.Equal((84UL, 37UL), p3.ButtonB);
                 Assert.Equal((10000000007870UL, 10000000006450UL), p3.Prize);
             },
             p1 => {
-                Assert.Equal((69, 23), p1.ButtonA);
-                Assert.Equal((27, 71), p1.ButtonB);
+                Assert.Equal((69UL, 23UL), p1.ButtonA);
+                Assert.Equal((27UL, 71UL), p1.ButtonB);
                 Assert.Equal((10000000018641UL, 10000000010279UL), p1.Prize);
             });
